fix(frontend): replace broken /insurance route with the working implementation
app/[locale]/insurance/page.tsx (the only reachable copy — middleware
redirects all non-prefixed paths to their [locale] equivalent) depended on
an entire shadcn/ui-style component kit that never existed in this repo:
components/ui/Tabs, components/ui/Card (capitalized; only a lowercase
card.tsx existed, missing CardDescription), and components/ui/{Button,
Input,Label,Select,Alert}. This broke the production build outright
(`Module not found`) — nobody had run `next build` before merging it.

There was a second, fully working implementation sitting at the
now-unreachable app/insurance/page.tsx (self-contained, built with plain
Tailwind markup and the real InsuranceDashboard/insuranceService/
riskAssessment modules, no missing dependencies). Moved it into the
reachable [locale] slot and deleted the broken duplicate — the
components/insurance/* trio and useInsuranceClaim hook were consumed by
nothing else.
